Rename README.md to SPOTIFY.md
What's impressive here is that popularity of one song doesn’t indicate most streamed infact if people are listening to all your music old and new that’s what gets you the streamed lewis capaldi just scraped a top 10 and he was the most streamed. In the table it shows the Uk appreciate albums as much as singles I would like to point out that if an album tends to have a popular song than the album tends to do well that always isn't the case though. The uk also seems to have a potty mouth or artist cant stop being explicit this may be a new trend in music. Sadly, as a lot of the songs in the top 50 weren't as popular but than again the popularity in this specific was it liked a feature in spotify was it most searched or listened too a bit more research should've been done in that aspect​